requirements.txt
twikit
supabase
python-dotenv
